web/requirements.txt
python-dotenv # 환경변수를 파일(.env)로 관리하는 라이브러리
streamlit # 웹 애플리케이션 개발 라이브러리
openai # 오픈AI API 클라이언트 라이브러리
jupyter # 데이터 분석 및 시각화 툴
pandas # 데이터 조작 및 분석 라이브러리
streamlit-markdown # 마크다운 문법을 사용하여 웹 애플리케이션 개발 라이브러리
groq # groq API 클라이언트 라이브러리
langchain # 언어 모델 체인 라이브러리
langchain_community # 언어 모델 체인 커뮤니티 라이브러리
langchain-ollama # 언어 모델 체인 올라마 라이브러리